+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 3.8.4
+- Changed the `makePurchase()` method signature when using parameters other than `product` – they have been extracted into `AdaptyPurchaseParameters`
+- [Android] Migrated to PBL v7
+- [Android] Added support for `obfuscatedProfileId` and `obfuscatedAccountId`
+- [iOS] Fixed support for Xcode 26
+
+# 3.8.3
+- [Android] Fixed AdaptyUI rendering issues
+
 # 3.8.2
 - [Android] Fixed AdaptyUI rendering issues